Scraper Tool: initial result pipeline framework
Initial framework for chaining together a series of functions where
failure may occur at any point, while allowing you to write pieces of
the pipeline assuming success up until that point.

In short, you write a function taking a value and either outputting a
value (if it can't failure) or a result object (if it can). Then you
decorate with @track(fmt), specifying the format of the function, to get
a new function which always takes a result and outputs a result.

See tests/test_result.py for examples of how to use the result pipeline
framework.